+++ lib/stardag/src/stardag/__init__.py
@@ -62,7 +62,12 @@ from stardag.exceptions import (
     StardagError,
     TokenExpiredError,
 )
-from stardag.polymorphic import Polymorphic, SubClass
+from stardag.polymorphic import (
+    NakedPolymorphicFieldError,
+    Polymorphic,
+    StrictPolymorphicTypeError,
+    SubClass,
+)
 from stardag.registry import registry_provider
 from stardag.target import (
     DirectoryTarget,
@@ -108,12 +113,14 @@ __all__ = [
     "LoadableTask",
     "LoadValidator",
     "LocalFileTarget",
+    "NakedPolymorphicFieldError",
     "namespace",
     "Polymorphic",
     "registry_provider",
     "StardagError",
     "StardagBaseModel",
     "StardagField",
+    "StrictPolymorphicTypeError",
     "SubClass",
     "Task",
     "TaskRehydrationError",